pyproject.toml
[project]
name = "yudou-collect"
version = "1.1.1"
description = "yudou-collect"
requires-python = ">=3.13"
dependencies = [
    "lxml>=6.0.2",
    "pycryptodome>=3.23.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "requests[socks]>=2.32.5",
    "tabulate>=0.9.0",
    "tenacity>=8.0.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.0",
    "types-requests>=2.32.4.20250809",
    "types-lxml>=2025.8.25",
    "types-tabulate>=0.9.0.20241207",
    "types-tqdm>=4.67.0.20250809",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"



